requirements.txt
appdirs==1.4.4
black==21.4b2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
idna==2.10
mypy-extensions==0.4.3
numpy==1.20.2
openai==0.6.3
pathspec==0.8.1
Pillow==8.2.0
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
toml==0.10.2
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
